requirements.txt
numpy~=2.2.3
colour-science==0.4.6
matplotlib>=3.10.0